Fix warnings for Win32 Compiler.cpp
This also involves regularising type for GlyphIDs to gid16 instead of
the mishmash of int and utf16 types currently used.
Set version fields to be uint32_t consisntently.

+++ compiler/Grammar/InputBuffer.cpp
@@ -73,7 +73,7 @@ void InputBuffer::fill(int amount)
 {
 	syncConsume();
 	// Fill the buffer sufficiently to hold needed characters
-	while (queue.entries() < amount + markerOffset) {
+	while (queue.entries() < size_t(amount + markerOffset)) {
 		// Append the next character
 		queue.append(getChar());
 	}
